+++ internal/assets/migrations/001_initial.sql
@@ -28,6 +28,7 @@ create table if not exists links (
 );
 insert into links (link) values ('mhrynenko/testapi');
 insert into links (link) values ('acstestapi');
+insert into links (link) values ('testapiacsdl');
 
 create index if not exists links_link_idx on links(link);
 
@@ -55,7 +56,7 @@ create table if not exists permissions (
     type text not null,
     created_at timestamp without time zone not null,
     expires_at timestamp without time zone not null,
-    updated_at timestamp without time zone not null default current_timestamp,
+    updated_at timestamp with time zone not null default current_timestamp,
     has_parent boolean not null default true,
     has_child boolean not null default false,
     parent_link text,