+++ internal/cli/main.go
@@ -8,8 +8,8 @@ import (
 	"github.com/alecthomas/kingpin"
 	"gitlab.com/distributed_lab/acs/github-module/internal/config"
 	"gitlab.com/distributed_lab/acs/github-module/internal/data"
+	"gitlab.com/distributed_lab/acs/github-module/internal/registrator"
 	"gitlab.com/distributed_lab/acs/github-module/internal/service"
-	"gitlab.com/distributed_lab/acs/github-module/internal/service/registrator"
 	"gitlab.com/distributed_lab/kit/kv"
 	"gitlab.com/distributed_lab/logan/v3"
 )
@@ -23,11 +23,10 @@ func Run(args []string) bool {
 	defer func() {
 		if rvr := recover(); rvr != nil {
 			log.WithRecover(rvr).Error("app panicked")
-			err := registrator.UnregisterModule(data.ModuleName, cfg.Registrator().OuterUrl)
+			err := registrator.NewRegistrar(cfg).UnregisterModule()
 			if err != nil {
 				log.WithError(err).Errorf("failed to unregister module %s", data.ModuleName)
 			}
-			log.Infof("unregistered module %s", data.ModuleName)
 		}
 	}()
 
@@ -41,12 +40,11 @@ func Run(args []string) bool {
 	go func() {
 		sig := <-signalChannel
 		log.Infof("service was interrupted by signal `%s`", sig.String())
-		err := registrator.UnregisterModule(data.ModuleName, cfg.Registrator().OuterUrl)
+		err := registrator.NewRegistrar(cfg).UnregisterModule()
 		if err != nil {
 			log.WithError(err).Errorf("failed to unregister module %s", data.ModuleName)
 			os.Exit(1)
 		}
-		log.Infof("unregistered module %s", data.ModuleName)
 		os.Exit(0)
 	}()