+++ internal/data/users.go
@@ -10,19 +10,16 @@ type Users interface {
 	New() Users
 
 	Upsert(user User) error
-	Delete(githubId int64) error
-
+	Delete() error
 	Select() ([]User, error)
 	Get() (*User, error)
 
 	Count() Users
 	GetTotalCount() (int64, error)
 
-	GetById(id int64) (*User, error)
-	GetByUsername(username string) (*User, error)
-	GetByGithubId(githubId int64) (*User, error)
-
 	FilterById(id *int64) Users
+	FilterByUsernames(usernames ...string) Users
+	FilterByGithubIds(githubIds ...int64) Users
 	FilterByLowerTime(time time.Time) Users
 	SearchBy(search string) Users
 
@@ -36,6 +33,7 @@ type User struct {
 	AvatarUrl string    `json:"avatar_url" db:"avatar_url" structs:"avatar_url"`
 	CreatedAt time.Time `json:"created_at" db:"created_at" structs:"-"`
 	UpdatedAt time.Time `json:"updated_at" db:"updated_at" structs:"-"`
+	Submodule *string   `json:"-" db:"-" structs:"-"`
 }
 
 type UnverifiedUser struct {