+++ internal/github/check_collaborator.go
@@ -10,7 +10,7 @@ import (
 	"gitlab.com/distributed_lab/logan/v3/errors"
 )
 
-func (g *github) CheckUserFromApi(link, username, typeTo string) (bool, *data.Permission, error) {
+func (g *github) CheckUserFromApi(link, username, typeTo string) (*CheckPermission, error) {
 	if typeTo == data.Repository {
 		return g.CheckRepoCollaborator(link, username)
 	}
@@ -18,13 +18,13 @@ func (g *github) CheckUserFromApi(link, username, typeTo string) (bool, *data.Pe
 		return g.CheckOrgCollaborator(link, username)
 	}
 
-	return false, nil, errors.Errorf("failed to check `%s` with `%s` type", link, typeTo)
+	return nil, errors.Errorf("failed to check `%s` with `%s` type", link, typeTo)
 }
 
-func (g *github) CheckRepoCollaborator(link, username string) (bool, *data.Permission, error) {
+func (g *github) CheckRepoCollaborator(link, username string) (*CheckPermission, error) {
 	req, err := http.NewRequest(http.MethodGet, fmt.Sprintf("https://api.github.com/repos/%s/collaborators/%s/permission", link, username), nil)
 	if err != nil {
-		return false, nil, errors.Wrap(err, " couldn't create request")
+		return nil, errors.Wrap(err, " couldn't create request")
 	}
 
 	req.Header.Set("Accept", "application/vnd.Github+json")
@@ -33,13 +33,13 @@ func (g *github) CheckRepoCollaborator(link, username string) (bool, *data.Permi
 
 	res, err := http.DefaultClient.Do(req)
 	if err != nil {
-		return false, nil, errors.Wrap(err, " error making http request")
+		return nil, errors.Wrap(err, " error making http request")
 	}
 
 	if res.StatusCode == http.StatusForbidden {
 		timeoutDuration, err := g.getDuration(res)
 		if err != nil {
-			return false, nil, errors.Wrap(err, "failed to get time duration from response")
+			return nil, errors.Wrap(err, "failed to get time duration from response")
 		}
 		g.log.Warnf("we need to wait `%d`", timeoutDuration)
 		time.Sleep(timeoutDuration)
@@ -47,11 +47,11 @@ func (g *github) CheckRepoCollaborator(link, username string) (bool, *data.Permi
 	}
 
 	if res.StatusCode == http.StatusNotFound {
-		return false, nil, nil
+		return nil, nil
 	}
 
 	if res.StatusCode != http.StatusOK {
-		return false, nil, errors.Errorf("unexpected status %s", res.Status)
+		return nil, errors.Errorf("unexpected status %s", res.Status)
 	}
 
 	returned := struct {
@@ -63,22 +63,24 @@ func (g *github) CheckRepoCollaborator(link, username string) (bool, *data.Permi
 	}{}
 
 	if err := json.NewDecoder(res.Body).Decode(&returned); err != nil {
-		return false, nil, errors.Wrap(err, " failed to unmarshal body")
+		return nil, errors.Wrap(err, " failed to unmarshal body")
 	}
 
-	return true, &data.Permission{
-		Link:        link,
-		Type:        data.Repository,
-		Username:    username,
-		GithubId:    returned.User.Id,
-		AccessLevel: returned.RoleName,
-	}, nil
+	return &CheckPermission{
+		true,
+		data.Permission{
+			Link:        link,
+			Type:        data.Repository,
+			Username:    username,
+			GithubId:    returned.User.Id,
+			AccessLevel: returned.RoleName,
+		}}, nil
 }
 
-func (g *github) CheckOrgCollaborator(link, username string) (bool, *data.Permission, error) {
+func (g *github) CheckOrgCollaborator(link, username string) (*CheckPermission, error) {
 	req, err := http.NewRequest(http.MethodGet, fmt.Sprintf("https://api.github.com/orgs/%s/memberships/%s", link, username), nil)
 	if err != nil {
-		return false, nil, errors.Wrap(err, " couldn't create request")
+		return nil, errors.Wrap(err, " couldn't create request")
 	}
 
 	req.Header.Set("Accept", "application/vnd.Github+json")
@@ -87,13 +89,13 @@ func (g *github) CheckOrgCollaborator(link, username string) (bool, *data.Permis
 
 	res, err := http.DefaultClient.Do(req)
 	if err != nil {
-		return false, nil, errors.Wrap(err, " error making http request")
+		return nil, errors.Wrap(err, " error making http request")
 	}
 
 	if res.StatusCode == http.StatusForbidden {
 		timeoutDuration, err := g.getDuration(res)
 		if err != nil {
-			return false, nil, errors.Wrap(err, "failed to get time duration from response")
+			return nil, errors.Wrap(err, "failed to get time duration from response")
 		}
 		g.log.Warnf("we need to wait `%d`", timeoutDuration)
 		time.Sleep(timeoutDuration)
@@ -101,11 +103,11 @@ func (g *github) CheckOrgCollaborator(link, username string) (bool, *data.Permis
 	}
 
 	if res.StatusCode == http.StatusNotFound {
-		return false, nil, nil
+		return nil, nil
 	}
 
 	if res.StatusCode != http.StatusOK {
-		return false, nil, errors.Errorf("unexpected status %s", res.Status)
+		return nil, errors.Errorf("unexpected status %s", res.Status)
 	}
 
 	returned := struct {
@@ -117,14 +119,16 @@ func (g *github) CheckOrgCollaborator(link, username string) (bool, *data.Permis
 	}{}
 
 	if err := json.NewDecoder(res.Body).Decode(&returned); err != nil {
-		return false, nil, errors.Wrap(err, " failed to unmarshal body")
+		return nil, errors.Wrap(err, " failed to unmarshal body")
 	}
 
-	return true, &data.Permission{
-		Link:        link,
-		Type:        data.Organization,
-		Username:    username,
-		GithubId:    returned.User.Id,
-		AccessLevel: returned.Role,
-	}, nil
+	return &CheckPermission{
+		true,
+		data.Permission{
+			Link:        link,
+			Type:        data.Organization,
+			Username:    username,
+			GithubId:    returned.User.Id,
+			AccessLevel: returned.Role,
+		}}, nil
 }