+++ internal/service/api/handlers/get_user_by_id.go
@@ -17,7 +17,7 @@ func GetUserById(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	user, err := UsersQ(r).GetById(userId)
+	user, err := UsersQ(r).FilterById(&userId).Get()
 	if err != nil {
 		Log(r).WithError(err).Errorf("failed to get user with id `%d`", userId)
 		ape.RenderErr(w, problems.InternalError())
@@ -28,6 +28,17 @@ func GetUserById(w http.ResponseWriter, r *http.Request) {
 		Log(r).Errorf("no user with id `%d`", userId)
 		ape.RenderErr(w, problems.NotFound())
 		return
+	}
+
+	permission, err := PermissionsQ(r).FilterByHasParent(false).FilterByParentLinks([]string{}...).FilterByGithubIds(user.GithubId).Get()
+	if err != nil {
+		Log(r).Errorf("failed to get submodule for user with id `%d`", userId)
+		ape.RenderErr(w, problems.NotFound())
+		return
+	}
+
+	if permission != nil {
+		user.Submodule = &permission.Link
 	}
 
 	ape.Render(w, models.NewUserResponse(*user))